chore: rename file to match function name
